Fetch!: Refactor store namespaces
