+++ internal/app/store/serverStore/memoryStore/memoryStore_test/hubRepository_internal_test.go
@@ -2,7 +2,7 @@ package memoryStore_test
 
 import (
 	"Chat/internal/app/model"
-	"Chat/internal/app/store/memoryStore"
+	"Chat/internal/app/store/serverStore/memoryStore"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
