+++ internal/app/store/serverStore/serverStore.go
@@ -1,0 +1,6 @@
+package serverStore
+
+// ServerStore storage server data
+type ServerStore interface {
+	Hub() HubRepository
+}
