--- internal/app/store/store.go
@@ -1,6 +1,0 @@
-package store
-
-// Store storage server data
-type Store interface {
-	Hub() HubRepository
-}
